cmake/compiler/definitions changes

+++ Synth/Source/audio/lfo/LFO.cpp
@@ -8,11 +8,11 @@
 namespace audio {
     LFO::LFO(std::atomic<float> *lfoValue, const std::atomic<float>* frequency) : lfoValue(lfoValue), frequency(frequency) {
     }
-    void LFO::shiftPhase(int numSamples, float sampleRate) {
+    void LFO::shiftPhase(int numSamples, double sampleRate) {
         phase += numSamples;
-        float samplesPerPhase = sampleRate / *frequency;
+        float samplesPerPhase = static_cast<float>(sampleRate) / *frequency;
         auto factor = (float)phase / samplesPerPhase;
-        auto value = std::sin((factor - (int)factor) * juce::MathConstants<float>::twoPi);
+        auto value = std::sin((factor - static_cast<float>(static_cast<int>(factor))) * juce::MathConstants<float>::twoPi);
         lfoValue->store(value);
     }
 }
