+++ Synth/Source/audio/lfo/LFO.h
@@ -6,7 +6,7 @@ namespace audio {
     class LFO {
     public:
         LFO(std::atomic<float>* lfoValue, const std::atomic<float>* frequency);
-        void shiftPhase(int numSamples, float sampleRate);
+        void shiftPhase(int numSamples, double sampleRate);
     private:
         long long phase = 0;
         std::atomic<float>* lfoValue;
